Fixed empty archive's name in legendastv provider
When logged in as a donor, there's an extra <a href=... class="seta">
</a> in the div, causing get_archives to return an archive with no name.
This will result in lots of unnecessary (and useless) downloads when
download_best_subtitles is called.

To fix we just specify the correct anchor (the one inside the <p>) in
get_archives.

(cherry picked from commit 854253a)