Revert changes related to multi episode video files
Allowing the Video.episode to be a list causes more problems than it fixes.
In case guessit returns a list of episodes, we only take the first (lowest) episode.
Subtitle providers also do not support a list of episodes. A multi episode subitle will always be listed under the lowest episode number.