Limit versions of rarfile that can be used
In markokr/rarfile#70 it's reported that the `custom_check` function
was removed from the rarfile pypi package. As a result the error
reported in Diaoul/subliminal#1013 occurs when a version of rarfile
newer than 3.1 is installed.

This fixes #1013 by limiting the range of rarfile versions.

Once the legendastv provider is changed to no longer use `custom_check`
this version constraint could be removed.